Remove logic CircularFlow from ConstraintHelper
